% Generated by roxygen2 (4.0.0): do not edit by hand
\name{disp.nbq}
\alias{disp.nbq}
\title{(private) Specify a NBQ dispersion model}
\usage{
disp.nbq(counts, eff.lib.sizes, x, phi.pre = 0.1, mu.lower = 1,
  mu.upper = Inf, pi.offset = median(pi.pre[subset, ]))
}
\arguments{
\item{counts}{a mxn matrix of NB counts}

\item{eff.lib.sizes}{a n-vector of estimated effective library sizes}

\item{x}{a nxp matrix, design matrix (specifying the treatment structure)}

\item{phi.pre}{a number, a preliminary constant dispersion value}

\item{mu.lower}{a number, rows with mu.pre < mu.lower will not be used for estimating the dispersion model}

\item{mu.upper}{a number, rows with mu.pre > mu.upper will not be used for estimating the dispersion model}

\item{pi.offset}{a scalar, an offset used in the NBQ model (see Details).}
}
\value{
a list

\item{fun}{a function that takes a vector, par, as
input and outputs a matrix of dispersion values (same dimension as
counts)}

\item{par.init}{a vector of length 3, initial values of par}
\item{lower}{a vector of length 3, lower bounds of the parameter values}
\item{upper}{a vector of length 3, upper bounds of the parameter values}
\item{subset}{a logical vector of length \eqn{m}, specifying the subset of rows to be used when estimating the dispersion model parameters.}
\item{pi.pre}{a m by n matrix, preliminary estimates of the relative frequencies.}
\item{pi.pre}{a m by n matrix, preliminary estimates of the relative frequencies.}
\item{pi.offset}{a scalar used as an offset in the NBQ model (see Details)}
}
\description{
Specify a NBQ dispersion model. The unknown parameters in the
specified model are to be estimated using the function
\code{optim.disp.pl} or \code{optim.disp.apl}.
}
\details{
Under this NBQ model, the dispersion is modeled as a quadratic
function of the preliminary estimates of the log mean realtive
frequencies (pi.pre):

log(phi) =  par[1] + par[2] * z + par[3] * z^2,

where z = log(pi.pre/pi.offset). By default, pi.offset is the median of pi.pre[subset,].
}

