% Generated by roxygen2 (4.0.0): do not edit by hand
\name{disp.nbs}
\alias{disp.nbs}
\title{(private) Specify a NBS dispersion model}
\usage{
disp.nbs(counts, eff.lib.sizes, x, df = 6, phi.pre = 0.1, mu.lower = 1,
  mu.upper = Inf)
}
\arguments{
\item{counts}{a mxn matrix of NB counts}

\item{eff.lib.sizes}{a n-vector of estimated effective library sizes}

\item{x}{a nxp matrix, design matrix (specifying the treatment structure)}

\item{df}{the number of interior nodes}

\item{phi.pre}{a number, a preliminary constant dispersion value}

\item{mu.lower}{a number, rows with mu.pre < mu.lower will not be used for estimating the dispersion model}

\item{mu.upper}{a number, rows with mu.pre > mu.upper will not be used for estimating the dispersion model}
}
\value{
a list

\item{fun}{a function that takes a vector, par, as input and
outputs a matrix (same dimension as \code{counts}) of dispersion
values. \code{par} will be used as linear-combination efficients
for the spline bases, the estimated dispersion values are a spline
function of log(pi.pre).  }

\item{par.init}{initial values of par}

\item{subset}{a logical vector of length \eqn{m}, specifying the
subset of genes to be used when estimating the model
parameters. Note that the estimated model will be applied to all
rows whenever possible, but only rows specified in \code{subset}
will be used to estimate the parameters of the dispersion model.}

\item{pi.pre}{a m by n matrix, preliminary estimates of the
relative frequencies.}

\item{s}{Basis matrix of the natural cubic spline evalued at the z = log(pi.pre)}
}
\description{
Specify a NBS dispersion model. The specified model are to be
estimated using the function \code{optim.disp.pl} or
\code{optim.disp.apl}.  Under this NBS model, the dispersion is
modeled as a smooth function (a natural cubic spline function) of
the preliminary estimates of the log mean realtive frequencies
(pi.pre).
}
\details{
\code{disp.nbs} calls the function \code{ns} to generate a set of
spline bases, using log(pi.pre) (converted to a vector) as the
predictor variable. Linear combinations of these spline bases are
smooth functions of log(pi.pre). The return value includes a
function, \code{fun}, to be optimized by \code{optim.disp.pl} or
\code{optim.disp.apl}. The parameter of that function is a vector
of linear combination coefficients of the spline bases.

\code{df+2} nodes are used when constructing the splint bases. The
Boundary.nodes are placed at the min and max values of
log(pi.pre). Two nodes are placed at the 0.05 and 0.95th quantiles
of log(pi.pre) and an additional \code{df-2} inner nodes are
equally spaced between the two nodes.

It is a challenging issue to determine the optimal number and
placement of the nodes.
}

