\name{disp.predictor.mu}
\alias{disp.predictor.mu}
\title{Dispersion precitor}
\usage{
disp.predictor.mu(nb.data, x, phi.pre = 0.1, mu.lower = 1, mu.upper = Inf)
}
\arguments{
  \item{nb.data}{}

  \item{x}{}

  \item{phi.pre}{}

  \item{mu.lower}{}

  \item{mu.upper}{}
}
\value{
a logical vector
}
\description{
Dispersion precitor
}

