% Generated by roxygen2 (4.0.0): do not edit by hand
\name{hist2d}
\alias{hist2d}
\title{2-d Histogram}
\usage{
hist2d(x, y, xlim = range(x), ylim = range(y), nbins)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector of the same length as x}

\item{xlim}{a vector of length 2, the range of x values}

\item{ylim}{a vector of length 2, the range of y values}

\item{nbins}{a single number giving the number of bins (the same for both x- and y- axes).}
}
\value{
a list
\item{x}{a vector of length nbins, the midpoints of each bin on the x-axis.}
\item{y}{a vector of length nbins, the midpoints of each bin on the y-axis.}
\item{z}{a \code{nbins} by \code{nbins} matrix of of counts. For each cell, the number of (x, y) inside.}
The list can be passed to \code{image()} directly for potting.
}
\description{
Commpute a 2-d histogram of the given data values. (not implemented yet: If \code{plot == TRUE}, plot the resulting histogram.)
}
\details{
This funciton divides the \code{xlim x ylim} region into \code{nbins x nbins}
equal-sized cells and count the number of (x,y) points in each cell.
}
\note{
Only points inside the region defined by \code{xlim x ylim}
(inclusive) will be counted. For each cell, the lower boundaries
are closed and upper boundaries are open.  A small number will be
added to the upper limits in xlim and ylim so that no points will
be on the region's upper boundaries.
}

