% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ma.plot}
\alias{ma.plot}
\title{(private) MA plot with differently expressed genes highlighted}
\usage{
ma.plot(test.out, top = NULL, q.cutoff = NULL, p.cutoff = NULL,
  col.sig = "magenta", main = "MA Plot", ...)
}
\arguments{
\item{test.out}{output from \code{\link{nbp.test}}}

\item{top}{a number indicating the number of genes to be declared
as differentially expressed}

\item{q.cutoff}{a number, q-value cutoff}

\item{p.cutoff}{a number, p-value cutoff}

\item{col.sig}{color}

\item{main}{label}

\item{...}{additional parameters to be passed to \code{\link{smart.plot}}}
}
\value{
a vector, indices of top genes.
}
\description{
Plot log (base 2) fold change vs average expression in RPM
(two-group pooled) (i.e., an MA plot) and highlight differentially
expressed genes on the plot.
}
\details{
Differentially expressed genes are those with smallest DE test
p-values. The user has three options to specify the set of DE
genes: the user can specify 1) the number of top genes to be
declared as significant;  2) a q-value cutoff; or 3) a p-value
cutoff.

The plot is based on the thinned counts. The units on the x-axis
is RPM (reads per million mapped reads). We use RPM so that the
results are more comparable between experiments with different
sequencing depth (and thus different column totals in the count
matrix).  We exclude rows (genes) with 0 total counts after
thinning.
}

