\name{mv.points}
\alias{mv.points}
\title{(private) Highlight a subset of points on the mean-variance plot}
\usage{
mv.points(counts, subset, ...)
}
\arguments{
  \item{counts}{a matrix of NB counts}

  \item{subset}{a numberic or logical vector indicating the
  subset}

  \item{...}{other

  of rows to be highlighted}
}
\description{
Highlight a subset of points on the mean-variance plot
}

