% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.nb.data}
\alias{plot.nb.data}
\title{Boxplot and scatterplot matrix of relative frequencies (after normalization)}
\usage{
\method{plot}{nb.data}(x, resolution = 50, hlim = 0.25, clip = 128,
  eps = 0.01, ...)
}
\arguments{
\item{x}{output from \code{\link{prepare.nb.data}}}

\item{resolution}{}

\item{hlim}{a single number controls the height of the bars in the}

\item{clip}{}

\item{eps}{a small positive number added to rpm}

\item{...}{currently not used
histograms}
}
\description{
Boxplot and scatterplot matrix of relative frequencies (after normalization)
}

