% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.nb.dispersion}
\alias{plot.nb.dispersion}
\title{Plot the estimated dispersion as a function of the
preliminarily estimated mean relative frequencies}
\usage{
\method{plot}{nb.dispersion}(x, ...)
}
\arguments{
\item{x}{output from \code{\link{estimate.dispersion}}}

\item{...}{additional parameters, currently unused}
}
\description{
Plot the estimated dispersion as a function of the
preliminarily estimated mean relative frequencies
}

