\name{plot.nbp}
\alias{plot.nbp}

\title{
  Diagnostic Plots for an NBP Object
}
\description{
  For output from \code{\link{nbp.test}}, produce a boxplot, an MA plot,
  a mean-variance plot and a mean-dispersion plot. On mean-variance and
  mean-dispersion plots, overlay curves corresponding to the estimated
  NBP model.
}
\usage{
 plot.nbp(x, ...);
}

\arguments{
  \item{x}{output from \code{\link{nbp.test}}.}
\item{...}{for future use.}
}
\author{
 Yanming Di <diy@stat.oregonstate.edu>
}

\seealso{
\code{\link{nbp.test}}
}

\examples{
## See the example for nbp.test
}
