\name{print.nb.dispersion}
\alias{print.nb.dispersion}
\title{Print the estimated dispersion model}
\usage{
\method{print}{nb.dispersion}(x, ...)
}
\arguments{
  \item{x}{output from from
  \code{\link{estimate.dispersion}}}

  \item{...}{additional parameters, currently unused}
}
\description{
Print the estimated dispersion model
}

