% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.nb.test}
\alias{print.nb.test}
\title{Print output from \code{\link{test.coefficient}}}
\usage{
\method{print}{nb.test}(x, ...)
}
\arguments{
\item{x}{output from \code{\link{test.coefficient}}}

\item{...}{currenty not used}
}
\description{
We simply print out the structure of \code{x}. (Currenlty the
method is equivalent to \code{print(str(x))}.)
}

