% Generated by roxygen2 (4.0.0): do not edit by hand
\name{thin.counts}
\alias{thin.counts}
\title{(private) Thin (downsample) counts to make the effective library
sizes equal.}
\usage{
thin.counts(y, current.lib.sizes = colSums(y),
  target.lib.sizes = min(current.lib.sizes))
}
\arguments{
\item{y}{an n by r matrix of counts}

\item{current.lib.sizes}{an r vector indicating current estimated library sizes}

\item{target.lib.sizes}{an r vector indicating target library sizes after thinning}
}
\value{
a list
\item{counts}{a matrix of thinned counts (same dimension as the input y).}
\item{librar.sizes}{library sizes after thinning, same as the input target.lib.sizes}
}
\description{
Thin (downsample) counts to make the effective library sizes equal.
}
\details{
The exact NB test for differential gene expression requires that
the effective library sizes (column sums of the count matrix
multiplied by normalization factors) are approximately equal.
This function will thin (downsample) the counts to make the
effective library sizes equal. Thinning may lose statistical
efficiency, but is unlikely to introduce bias. The reason to use
thinning, not scaling, is because Poisson counts after thinning
are still Poisson, but Poisson counts after scaling will not be
Poisson.
}

