% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{NBPSeq-package}
\alias{NBPSeq}
\alias{NBPSeq-package}
\title{Negative Binomial Regression Models for Statistical Analysis of RNA-Sequencing Data}
\description{
Negative binomial (NB) two-group and regression models for
RNA-Sequencing data analysis.
}
\details{
See the examples of \code{\link{test.coefficient}} and
\code{\link{exact.nb.test}} for typical workflows of using this package.
}

