\name{plot.nb.data}
\alias{plot.nb.data}
\title{Boxplot and scatterplot matrix of relative frequencies (after normalization)}
\usage{
\method{plot}{nb.data}(x, ...)
}
\arguments{
  \item{x}{output from \code{\link{prepare.nb.data}}}

  \item{...}{currently not used}
}
\description{
Boxplot and scatterplot matrix of relative frequencies
(after normalization)
}

